Slaved away at the computer until I got ADD, EDIT, SEARCH functioning completely
Went through all the components composing Discount and adjusted it to
work with Products. Everything seems to function, I've added items,
tried adding the same item again and an error message displays, I've
changed an item name. I haven't used the search item since the inventory
is so small. But I think we are good.